chore: rework contests archi
